requirements.txt
Pyrogram>=2.0.59
TgCrypto==1.2.4
python-dotenv 
speedtest-cli
telegraph[aio]
pyrate_limiter 
apscheduler
cachetools 
aiofiles
psutil
Pillow
httpx
bs4
uvloop
